Inital Data Fuction

+++ filmAdvice/movie/models.py
@@ -2,8 +2,8 @@ from django.db import models
 
 
 class Movie(models.Model):
-    movie_id = models.DecimalField('Film ID', max_digits=10, decimal_places=8, default=0)
-    imdb_id = models.DecimalField('IMDB ID', max_digits=10, decimal_places=8, default=0)
+    movie_id = models.PositiveIntegerField('Film ID', default=0)
+    imdb_id = models.PositiveIntegerField('IMDB ID', default=0)
     movie_name = models.CharField('Film Adı', max_length=150, default="")
     slug = models.SlugField(null=True)
     created_at = models.DateTimeField('Kayıt Tarihi', auto_now_add=True, editable=False, null=True, blank=True)
